refactor: remove some demo stuff
This is only demo stuff. I understand that at the beginning it was
easier to put everything at the same place but now it is too big and
needs more structuring.

So if you are creating a presentation with a demo you want to unit test,
please put the demo and its tests into your repository (for instance in
the same repository where you put the presentation).